requirements.txt
Django>=4.2.5,<4.3
djangorestframework>=3.14,<3.15
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.26.4,<0.26.5
Pillow>=8.2.0,<8.3.0
